requirements1.txt
absl-py==1.0.0
astor==0.8.1
beautifulsoup4==4.10.0
cached-property==1.5.2
certifi==2021.10.8
chardet==3.0.4
click==7.1.2
faiss-cpu==1.7.1.post3
filelock==3.4.0
Flask==1.1.4
gast==0.2.2
gdown==4.2.0
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
importlib-metadata==4.8.3
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.0.1
mtcnn==0.1.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python-headless==4.5.4.58
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.21.0
requests-oauthlib==1.0.0
scipy==1.5.4
six==1.16.0
soupsieve==2.3.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.24.3
Werkzeug==1.0.1
wrapt==1.13.3
zipp==3.6.0
